# Fix clinic activity logs query performance-created-by-agentic