--- a/src/main/resources/db/migration/V20250528_1__add_numeric_value_index.sql
+++ b/src/main/resources/db/migration/V20250528_1__add_numeric_value_index.sql
@@ -1,0 +1,2 @@
+-- Add index to improve query performance on numeric_value column
+CREATE INDEX IF NOT EXISTS idx_clinic_activity_logs_numeric_value ON clinic_activity_logs(numeric_value);