This PR addresses a critical performance issue in the clinic activity logs query.

Problem:
- Query execution time increased from 592.55μs to 4.69s (8000x degradation)
- Full sequential scan on clinic_activity_logs table
- Missing index on numeric_value column causing performance degradation

Solution:
- Added index on numeric_value column
- Verified performance improvement:
  - Before: 4.69s execution time
  - After: 0.334ms execution time (14,000x improvement)

Validation:
- Index creation successful
- Query now uses index scan instead of sequential scan
- Performance returned to expected levels

Related:
- Issue ID: 5aa1fb7a-3656-11f0-8db3-ca59c7e8e81d
- Endpoint: /api/clinic-activity/query-logs